+++ src/main/java/kr/co/conceptbe/auth/domain/authcode/AuthCodeRequestUrlProvider.java
@@ -1,6 +1,6 @@
 package kr.co.conceptbe.auth.domain.authcode;
 
-import kr.co.conceptbe.member.OauthServerType;
+import kr.co.conceptbe.member.domain.OauthServerType;
 
 public interface AuthCodeRequestUrlProvider {
