+++ src/main/java/kr/co/conceptbe/auth/domain/authcode/AuthCodeRequestUrlProviderHandler.java
@@ -6,7 +6,7 @@ import static java.util.stream.Collectors.toMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import kr.co.conceptbe.member.OauthServerType;
+import kr.co.conceptbe.member.domain.OauthServerType;
 import kr.co.conceptbe.member.exception.NotFoundOauthServerTypeException;
 import org.springframework.stereotype.Component;
