+++ src/main/java/kr/co/conceptbe/auth/domain/client/OauthMemberClientHandler.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import kr.co.conceptbe.auth.infra.oauth.dto.OauthMemberInformation;
-import kr.co.conceptbe.member.OauthServerType;
+import kr.co.conceptbe.member.domain.OauthServerType;
 import kr.co.conceptbe.member.exception.NotFoundOauthServerTypeException;
 import org.springframework.stereotype.Component;
