+++ src/main/java/kr/co/conceptbe/auth/infra/oauth/dto/OauthMemberInformation.java
@@ -1,6 +1,6 @@
 package kr.co.conceptbe.auth.infra.oauth.dto;
 
-import kr.co.conceptbe.member.OauthServerType;
+import kr.co.conceptbe.member.domain.OauthServerType;
 
 public record OauthMemberInformation(
     String oauthId,
