+++ src/main/java/kr/co/conceptbe/auth/infra/oauth/kakao/authcode/KakaoAuthCodeRequestUrlProvider.java
@@ -2,7 +2,7 @@ package kr.co.conceptbe.auth.infra.oauth.kakao.authcode;
 
 import kr.co.conceptbe.auth.domain.authcode.AuthCodeRequestUrlProvider;
 import kr.co.conceptbe.auth.infra.oauth.kakao.config.KakaoOauthConfig;
-import kr.co.conceptbe.member.OauthServerType;
+import kr.co.conceptbe.member.domain.OauthServerType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
