+++ src/main/java/kr/co/conceptbe/auth/presentation/OauthController.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 import kr.co.conceptbe.auth.application.OauthService;
 import kr.co.conceptbe.auth.application.dto.OauthMemberResponse;
 import kr.co.conceptbe.auth.presentation.dto.TokenResponse;
-import kr.co.conceptbe.member.OauthServerType;
+import kr.co.conceptbe.member.domain.OauthServerType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
