+++ src/main/java/kr/co/conceptbe/auth/presentation/OauthServerTypeConverter.java
@@ -1,6 +1,6 @@
 package kr.co.conceptbe.auth.presentation;
 
-import kr.co.conceptbe.member.OauthServerType;
+import kr.co.conceptbe.member.domain.OauthServerType;
 import org.springframework.core.convert.converter.Converter;
 
 public class OauthServerTypeConverter implements Converter<String, OauthServerType> {
