+++ src/main/java/kr/co/conceptbe/comment/Comment.java
@@ -14,9 +14,8 @@ import java.util.ArrayList;
 import java.util.List;
 import kr.co.conceptbe.common.entity.base.BaseTimeEntity;
 import kr.co.conceptbe.idea.domain.Idea;
-import kr.co.conceptbe.member.Member;
+import kr.co.conceptbe.member.domain.Member;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.OnDelete;
