+++ src/main/java/kr/co/conceptbe/idea/presentation/IdeaController.java
@@ -10,7 +10,7 @@ import kr.co.conceptbe.idea.dto.IdeaDetailResponse;
 import kr.co.conceptbe.idea.presentation.dto.request.IdeaRequest;
 import kr.co.conceptbe.idea.presentation.dto.response.BestIdeaResponse;
 import kr.co.conceptbe.idea.presentation.dto.response.IdeaResponse;
-import kr.co.conceptbe.member.Member;
+import kr.co.conceptbe.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
