+++ src/main/java/kr/co/conceptbe/member/application/MemberService.java
@@ -4,10 +4,10 @@ import java.util.List;
 import kr.co.conceptbe.auth.presentation.dto.TokenResponse;
 import kr.co.conceptbe.auth.support.JwtProvider;
 import kr.co.conceptbe.common.entity.domain.persistence.PurposeRepository;
-import kr.co.conceptbe.member.Member;
-import kr.co.conceptbe.member.MemberPurpose;
-import kr.co.conceptbe.member.MemberRepository;
-import kr.co.conceptbe.member.MemberSkillCategory;
+import kr.co.conceptbe.member.domain.Member;
+import kr.co.conceptbe.member.domain.MemberPurpose;
+import kr.co.conceptbe.member.persistence.MemberRepository;
+import kr.co.conceptbe.member.domain.MemberSkillCategory;
 import kr.co.conceptbe.member.application.dto.DetailSkillResponse;
 import kr.co.conceptbe.member.application.dto.FindSignUpResponse;
 import kr.co.conceptbe.member.application.dto.MainSkillResponse;
