+++ src/test/java/kr/co/conceptbe/member/application/MemberServiceTest.java
@@ -7,8 +7,8 @@ import java.util.List;
 import kr.co.conceptbe.auth.presentation.dto.TokenResponse;
 import kr.co.conceptbe.auth.support.JwtProvider;
 import kr.co.conceptbe.common.entity.domain.persistence.PurposeRepository;
-import kr.co.conceptbe.member.Member;
-import kr.co.conceptbe.member.MemberRepository;
+import kr.co.conceptbe.member.domain.Member;
+import kr.co.conceptbe.member.persistence.MemberRepository;
 import kr.co.conceptbe.member.application.dto.SignUpRequest;
 import kr.co.conceptbe.member.application.dto.SignUpSkillRequest;
 import kr.co.conceptbe.member.fixture.MemberFixture;
